E2E, test_dataset_tail(): file renaming
