+++ src/e2e/app/cli/repo-tests/src/rest_api/mod.rs
@@ -7,8 +7,8 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
+mod test_dataset;
 mod test_openapi;
-mod test_rest_api;
 
+pub use test_dataset::*;
 pub use test_openapi::*;
-pub use test_rest_api::*;
